requirements.txt
bcrypt==4.1.2
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
